requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.20
pandas==2.2.2
python-dateutil==2.8.2
spacy==3.6.0
nltk==3.8.1
reportlab==4.1.0
streamlit==1.27.0
plotly==5.15.0
requests==2.31.0
